+++ src/components/folderList/FolderData.jsx
@@ -1,8 +1,15 @@
 import React from "react";
 import "./FolderData.css";
 
-function FolderData({ name }) {
-  return <li className="folderData">{name}</li>;
+function FolderData({ name, onClick, selected }) {
+  return (
+    <li
+      className={`folderData ${selected ? "selected" : ""}`}
+      onClick={onClick}
+    >
+      {name}
+    </li>
+  );
 }
 
 export default FolderData;
