+++ src/components/footer/Footer.css
@@ -9,7 +9,6 @@
 /* 하단 바 */
 .footer-var {
   padding-top: 60px;
-  position: relative;
 }
 
 .bottom-nav {
